requirements.txt
Django>=5.1
click>=8.1
typer>=0.12
rich>=13.7
docker>=7.0
fastapi>=0.115
uvicorn>=0.30
nats-py>=2.8
httpx>=0.27
